+++ sync-for-payroll/src/codatsyncpayroll/models/operations/get_create_accounts_model.py
@@ -12,7 +12,9 @@ from typing import Optional
 @dataclasses.dataclass
 class GetCreateAccountsModelRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a connection."""
     
 
 
@@ -21,11 +23,14 @@ class GetCreateAccountsModelRequest:
 @dataclasses.dataclass
 class GetCreateAccountsModelResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     push_option: Optional[shared_pushoption.PushOption] = dataclasses.field(default=None)
     r"""OK"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""