+++ sync-for-payroll/src/codatsyncpayroll/models/shared/pushfieldvalidation.py
@@ -12,7 +12,10 @@ from typing import Optional
 @dataclasses.dataclass
 class PushFieldValidation:
     details: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('details') }})
+    r"""Details on the validation issue."""
     field: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('field'), 'exclude': lambda f: f is None }})
+    r"""Field name that resulted in the validation issue."""
     ref: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('ref'), 'exclude': lambda f: f is None }})
+    r"""Unique reference identifier for the validation issue."""